+++ src/otx/engine/utils/hpo/__init__.py
@@ -3,7 +3,7 @@
 #
 """Functions and Classes to run HPO in the engine."""
 
-from .hpo import execute_hpo
+from .hpo_api import execute_hpo
 from .hpo_trial import update_hyper_parameter
 
 __all__ = ["execute_hpo", "update_hyper_parameter"]
